+++ src/Client/EventLoop/Actions.hs
@@ -10,12 +10,16 @@ Maintainer  : emertens@gmail.com
 
 module Client.EventLoop.Actions
   ( Action(..)
-  , ActionMap
+  , KeyMap
   , keyToAction
   , initialKeyMap
+  , addKeyBinding
+  , keyMapEntries
+  , parseKey
   ) where
 
 import           Graphics.Vty.Input.Events
+import           Config.Schema.Spec
 import           Data.List
 import           Data.Map (Map)
 import qualified Data.Map as Map
@@ -23,7 +27,9 @@ import           Data.HashMap.Lazy (HashMap)
 import qualified Data.HashMap.Lazy as HashMap
 import           Data.Text (Text)
 import qualified Data.Text as Text
+import           Text.Read
 
+-- | Actions that can be invoked using the keyboard.
 data Action
 
   = ActDelete
@@ -72,8 +78,21 @@ data Action
   | ActToggleHideMeta
   deriving (Eq, Ord, Read, Show)
 
-newtype ActionMap = ActionMap (Map [Modifier] (Map Key Action))
+-- | Lookup table for keyboard events to actions. Use with
+-- keyToAction.
+newtype KeyMap = KeyMap (Map [Modifier] (Map Key Action))
   deriving (Show)
+
+keyMapEntries :: KeyMap -> [([Modifier], Key, Action)]
+keyMapEntries (KeyMap m) =
+  [ (mods, k, a)
+     | (mods, m1) <- Map.toList m
+     , (k,a) <- Map.toList m1
+     ]
+
+instance Spec Action where
+  valuesSpec = customSpec "action" anyAtomSpec $ \a ->
+                fst <$> HashMap.lookup a actionInfos
 
 
 -- | Names and default key bindings for each action.
@@ -135,26 +154,64 @@ actionInfos =
   ]
 
 
--- | Lookup the correct action to perform in response to a particular key event.
+-- | Lookup the action to perform in response to a particular key event.
 keyToAction ::
-  ActionMap  {- ^ actions      -} ->
+  KeyMap     {- ^ actions      -} ->
   Text       {- ^ window names -} ->
   [Modifier] {- ^ key modifier -} ->
   Key        {- ^ key          -} ->
   Action     {- ^ action       -}
 keyToAction _ names [MMeta] (KChar c)
   | Just i <- Text.findIndex (c==) names = ActJump i
-keyToAction (ActionMap m) names modifier key =
+keyToAction (KeyMap m) names modifier key =
   case Map.lookup key =<< Map.lookup modifier m of
     Just a -> a
     Nothing | KChar c <- key, null modifier -> ActInsert c
             | otherwise                     -> ActIgnored
 
+
+-- | Bind a keypress event to a new action.
+addKeyBinding ::
+  [Modifier] {- ^ modifiers -} ->
+  Key        {- ^ key       -} ->
+  Action     {- ^ action    -} ->
+  KeyMap     {- ^ actions   -} ->
+  KeyMap
+addKeyBinding mods k a (KeyMap m) = KeyMap $
+  Map.alter (Just . maybe (Map.singleton k a) (Map.insert k a)) mods m
+
+
 -- | Default key bindings
-initialKeyMap :: ActionMap
-initialKeyMap = ActionMap $
+initialKeyMap :: KeyMap
+initialKeyMap = KeyMap $
   Map.fromListWith Map.union
     [ (mods, Map.singleton k act)
       | (act, mks) <- HashMap.elems actionInfos
       , (mods, k)  <- mks
       ]
+
+
+parseKey :: String -> Maybe ([Modifier], Key)
+parseKey [c] = Just ([], KChar c)
+
+parseKey ('F':xs) =
+  do i <- readMaybe xs
+     Just ([], KFun i)
+
+parseKey ('C':'-':xs) =
+  do (m,k) <- parseKey xs
+     Just (MCtrl:m, k)
+
+parseKey ('M':'-':xs) =
+  do (m,k) <- parseKey xs
+     Just (MMeta:m, k)
+
+parseKey ('A':'-':xs) =
+  do (m,k) <- parseKey xs
+     Just (MAlt:m, k)
+
+parseKey ('S':'-':xs) =
+  do (m,k) <- parseKey xs
+     Just (MShift:m, k)
+
+parseKey _ = Nothing