+++ src/Client/State/Focus.hs
@@ -51,6 +51,7 @@ data Subfocus
   | FocusPalette     -- ^ Show current palette
   | FocusMentions    -- ^ Show all mentions
   | FocusDigraphs    -- ^ Show all digraphs
+  | FocusKeyMap      -- ^ Show key bindings
   | FocusHelp (Maybe Text) -- ^ Show help window with optional command
   deriving (Eq,Show)